Initial commit with Identity and Controller
This initial commit for csi-scaleio implements basic functionality for
the Identity and Controller services. This does not implement any of the
Node service functionality, nor does it implement idempotency.

This allows the storage plugin to be run with
X_CSI_SCALEIO_CONTROLLERONLY=true and then to do controller commands of
ls, attach/detach, create/delete.